+++ pjrpc/server/integration/django/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class JsonRPCConfig(AppConfig):
+
+    name = 'pjrpc.server.integration.django'
+    verbose_name = "pjrpc application"