+++ pjrpc/server/utils.py
@@ -52,8 +52,8 @@ def remove_suffix(s: str, suffix: str) -> str:
         return s
 
 
-def join_path(*paths) -> str:
-    result = ''
+def join_path(path, *paths) -> str:
+    result = path
     for path in paths:
         result = f'{result.rstrip("/")}/{path.lstrip("/")}'